refactor: Square 컴포넌트의 useTouchEndListener와 handleDropCharacter를 별도 파일로 분리
