+++ frontend/src/hooks/useTouchEndListener.ts
@@ -1,0 +1,38 @@
+import { useEffect } from 'react';
+
+export const useTouchEndListener = (
+  row: number,
+  col: number,
+  handleDrop: Function
+) => {
+  useEffect(() => {
+    // 터치가 끝났을 때 실행되는 이벤트 핸들러
+    const handleTouchEnd = () => {
+      const bounds = document
+        .querySelector(`.square.row-${row}.col-${col}`) // 해당 square의 DOM 요소를 선택
+        ?.getBoundingClientRect(); // 해당 square의 크기와 위치를 가져오기
+
+      const touchingPosition = (window as any).touchingPosition; // 현재 터치 중인 위치 정보
+
+      // square 영역 내에 있으면 드롭 처리
+      if (
+        bounds &&
+        touchingPosition &&
+        touchingPosition.x >= bounds.left &&
+        touchingPosition.x <= bounds.right &&
+        touchingPosition.y >= bounds.top &&
+        touchingPosition.y <= bounds.bottom
+      ) {
+        handleDrop();
+      }
+    };
+
+    // 터치 이벤트 리스너 등록
+    window.addEventListener('character-touch-end', handleTouchEnd); // `character-touch-end` 이벤트 발생 시 handleTouchEnd 함수 실행
+
+    // 컴포넌트가 언마운트될 때 이벤트 리스너 제거
+    return () => {
+      window.removeEventListener('character-touch-end', handleTouchEnd);
+    };
+  }, [row, col, handleDrop]);
+};
